Config working like it was before the change to org mode.
This latest addition includes adding babel languages. Most languages work, but for some reason javascript does not work. The issue is that the path in fish does not transfer. Even after adding the path of node, there is an error about printing (even when I'm not printing anything). I guess it's the way that node is called.

I'm not planning on updating this now to work with javascript.